requirements.txt
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.0.3
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
six==1.13.0
redis==4.2.2
